requirements.txt
flask
torch
torchvision
timm
Pillow
numpy<2.0
requests